Separate the code a bit
